First switch to nextjs
