+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["@babel/preset-react"]
+  "presets": ["next/babel"],
+  "plugins": [["styled-components", { "ssr": true }]]
 }
