+++ package.json
@@ -9,7 +9,7 @@
   "private": true,
   "scripts": {
     "build": "webpack --mode production",
-    "dev": "webpack-dev-server -d --open",
+    "dev": "next dev",
     "lint": "eslint --ext .ts --ext .tsx --ext .js src/**",
     "test": "jest",
     "test-ci": "jest --ci --runInBand --reporters=default --reporters=jest-junit"
@@ -23,6 +23,7 @@
     "@types/react-dom": "^16.8.4",
     "@types/styled-components": "^4.1.16",
     "babel-loader": "^8.0.6",
+    "babel-plugin-styled-components": "^1.12.0",
     "css-loader": "^2.1.1",
     "dotenv": "^8.0.0",
     "enzyme": "^3.10.0",
@@ -78,5 +79,8 @@
   "jest-junit": {
     "outputDirectory": "./test-results/jest",
     "outputName": "./results.xml"
+  },
+  "dependencies": {
+    "next": "^10.0.6"
   }
 }
