+++ src/components/PersonalProject.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { IStyledProps } from '../pages/App';
+import { IStyledProps } from '../utils/theme';
 
 export interface IPersonalProject {
   name: string;
