+++ src/components/Project.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { IStyledProps } from '../pages/App';
+import { IStyledProps } from '../utils/theme';
 
 interface IProjectProps {
   title: string;
